Add missing include to pg_upgrade/version.c
Commit 8d48e6a724 uses RELKIND_ constants when building the query, but
did not include the header defining them. On 10+ this header is already
included, but on 9.6 and earlier it was missing. It compiles just fine,
but then fails during execution

  ERROR:  column "relkind_relation" does not exist

Fix by adding the necessary header file, and backpatch to 9.4-.

Backpatch-to: 9.4-
Discussion: https://postgr.es/m/16045-673e8fa6b5ace196%40postgresql.org