First-draft release notes for 12.1.
As usual, the release notes for other branches will be made by cutting
these down, but put them up for community review first.  Note that a
fair percentage of the entries apply only to prior branches because
their issue was already fixed in 12.0.  I'll remove those from the 12.1
list later.