Revert recent change of ScanOption values and renumber SO_TYPE_TIDSCAN.
Commit 598b466e80 in v12 renumbered ScanOption enum values to add
the option for Tid scan. But the change of the codes assigned to
the existing values caused an ABI break and may break some extensions
depending on them. This should be avoided in minor version.

This commit reverts the renumbering of ScanOption enum values made by
commit 598b466e80 in v12 and put the ScanOption for Tid scan with new value.
This is applied only to v12.

Per complaint from Tom Lane.

Discussion: https://postgr.es/m/5261.1581103527@sss.pgh.pa.us