pg_query: Track comments separately from whitespace in lexer/parser
For syntax highlighting and extracting comments from a query, its very
helpful to know the exact locations of a comment in the query string.

Previously the lexer discarded all comments as whitespace, making it
impossible to determine where they are located in the query string. With
this change, the lexer returns them as SQL_COMMENT/C_COMMENT tokens.