Fix the bug when create schema authentication rolname.
When we create schema authentication rolname create
table/sequence/view sch.obj and sch is not NULL. In
parse_utilcmd.c we will check whether schemaName and
rolname are same or not. This will cause the database
core due to strcmp will NULL pointer.

Here, we make CREATE SCHEMA AUTHENTICATION rolname is
the same as CREATE SCHEMA rolname AUTHENTICATION rolname.