XLog: replicate fork file and encrypt tuples (#59)
1. Inserts and Updates now are encrypted in WAL.
We encrypt new tuples directly in Buffer after they were insrerted there. To
pass it to XLog we could memcpy Buffer data into into the tuple. But later
tuple has to be unencrypted for index instertions etc. So we pass directly
data from the Buffer into XLog.

2. Log into WAL and replicate *.tde forks creation.

3. Added docker-compose for the streaming replication test setup.
(not perfect - needs two `up -d` in a row to start the secondary)

4. Added tests for multi inserts. Need tests for replications though.